// mobile/src/variables/enum.ts
export enum EStorage {
  ACCESS_TOKEN = 'ACCESS_TOKEN',
  CLIENT_ID = 'CLIENT_ID',
}

export enum EClientStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  PENDING = 'PENDING',
}

export enum EInputType {
  TEXT = 'TEXT',
  TEXTAREA = 'TEXTAREA',
  PASSWORD = 'PASSWORD',
  PHONE_NUMBER = 'PHONE_NUMBER',
}

export enum EHomeworkType {
  ACTIVITY = 'Activity',
  QUESTIONNAIRE = 'Questionnaire',
  WRITTEN_TASK = 'WrittenTask',
  VIDEO = 'Video',
}

export enum EFeeling {
  ANGRY = 1,
  VERY_SAD = 2,
  SAD = 3,
  HAPPY = 4,
  SMILE = 5,
}